requirements.txt
selenium>=4.18.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
